PythonAPI/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
fastapi==0.115.8
h11==0.14.0
idna==3.10
pydantic==2.10.6
pydantic_core==2.27.2
python-nmap==0.7.1
scapy==2.6.1
sniffio==1.3.1
speedtest==0.0.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
fastapi==0.115.8
uvicorn==0.34.0
python-nmap==0.7.1
scapy==2.6.1
python-multipart==0.0.9